+++ example/lib/core/constants/app_data.dart
@@ -26,7 +26,7 @@ class AppData {
   // Version of the WEB build, usually same as package, but it also has a
   // build numbers.
   static const String versionMajor = '3';
-  static const String versionMinor = '0';
+  static const String versionMinor = '1';
   static const String versionPatch = '0';
   static const String versionBuild = '01';
   static const String version = '$versionMajor.$versionMinor.$versionPatch '